Develop

+++ database/migrations/2023_06_19_122358_create_rating_and_comments_table.php
@@ -18,7 +18,7 @@ return new class extends Migration
             $table->foreignId('product_id')->constrained('products');
             $table->foreignId('user_id')->constrained('users');
             $table->text('message')->nullable();
-            $table->unsignedSmallInteger('rating');
+            $table->unsignedTinyInteger('rating');
             $table->unique(['product_id', 'user_id']);
             $table->timestamps();
         });
